refactor: put angle equation in angle
